pyproject.toml
[tool.poetry]
name = "android"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
django-ninja = "^0.22.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
